Break out the startScreen class to its own file: StartScreen.js
